requirements-gpu.txt
ipython==7.2.0
matplotlib==3.0.2
numpy==1.16.0
tensorflow-gpu==1.12.0
pandas==0.24.0
scikit-learn==0.20.3